+++ CHANGELOG.md
@@ -157,6 +157,7 @@ This patch release contains only minor dependency bumps.
 * (x/group) [#25920](https://github.com/cosmos/cosmos-sdk/pull/25920) Expand voting period check to verify period is positive instead of nonzero.
 * (types/address) [#25944] (https://github.com/cosmos/cosmos-sdk/pull/25944) correct sort comparator in Compose to satisfy strict weak ordering.
 * (baseapp) [#26063](https://github.com/cosmos/cosmos-sdk/pull/26063) Fixes an issue where values embedded in context during ante handling were wiped after the handlers returned.
+* (store/iavl) [#25914](https://github.com/cosmos/cosmos-sdk/pull/25914) fix(store/iavl): use versioned store for /subspace historical queries.
 
 ### Deprecated